requirements.txt
python-telegram-bot
apscheduler
requests
python-dotenv
google-cloud-translate